opencc-converter/requirements.txt
# opencc-converter/requirements.txt
opencc-python-reimplemented==0.1.7
requests==2.31.0
pyyaml==6.0.1
colorama==0.4.6